修复timer是64bit位宽不对的问题

+++ include/cpu.h
@@ -17,9 +17,9 @@ typedef struct CPU {
     uint32_t prev_regs[14];     // 上一周期（或上一条指令执行前）寄存器快照：用于比较变化、生成display信息、触发trace/观察点，避免在同一周期内读写竞争；R0-R13按顺序对应
     struct BUS bus;             // CPU connected to BUS
     uint8_t  domain;
-    uint32_t timer[2];
+    uint64_t timer[2];
     uint8_t  timer_enabled[2];
-    uint32_t timer_threshold[2];   // 计时器阈值，当计数达到该值触发跳转
+    uint64_t timer_threshold[2];   // 计时器阈值，当计数达到该值触发跳转
     uint32_t timer_target_pc[2];   // 计时器触发后的目标PC（绝对地址，DRAM 基址空间）
 } CPU;
